wind_correlation_analysis/requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
matplotlib>=3.4.0
pyyaml>=5.4.0

# Data acquisition
requests>=2.26.0
cdsapi>=0.5.1  # For ERA5 downloads (requires account setup)

# Data processing
xarray>=0.19.0
netCDF4>=1.5.7

# Optional: Jupyter for interactive analysis
jupyter>=1.0.0
ipython>=7.27.0

# Optional: Additional visualization
seaborn>=0.11.0

# Testing (optional)
pytest>=6.2.0
pytest-cov>=2.12.0
